+++ src/js/playground-interactive.ts
@@ -1,86 +1,147 @@
-/*
- * File for main.html - Currently useless
+/**
+ * @summary Main file for playground-interactive.ts, which implements the compilation and online editor behaviour.
  */
 
-// main items
-const codeInput: HTMLTextAreaElement = document.querySelector("#code-editor-textarea");
-const codeInputResult: HTMLElement = document.querySelector("#highlighting-field-content");
+const localStorageIdentifier = "kipper-code-editor-content";
+
+// Editor elements
+const codeEditor: HTMLDivElement  = document.querySelector("#code-editor");
+const codeTextArea: HTMLTextAreaElement = document.querySelector("#code-editor-textarea");
+const codeTextAreaResultWrapper: HTMLElement = document.querySelector("#highlighting-field");
+const codeTextAreaResult: HTMLElement = document.querySelector("#highlighting-field-content");
 const textSavingState: HTMLDivElement = document.querySelector("#text-saving-state");
 
-// menu buttons
+// Menu buttons
 const runCodeListItem: HTMLLIElement = document.querySelector("#run-code-list-item");
 let runCodeButton: HTMLButtonElement = document.querySelector("#run-code-list-item button");
 const copyCodeButton: HTMLButtonElement = document.querySelector("#copy-code-list-item button");
 const clearContentButton: HTMLButtonElement = document.querySelector("#clear-content-list-item button");
 
-// for now, don't update the side-editor
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const sideEditor: HTMLDivElement = document.querySelector("#side-editor");
+// Sidebar editor fields
+const shellOutput: HTMLDivElement = document.querySelector("#shell-output");
+const shellOutputResult: HTMLElement = document.querySelector("#shell-sidebar-highlight-field-content");
 
-// sidebar buttons
+// Sidebar buttons
 const consoleOutputButton: HTMLButtonElement = document.querySelector("#console-output-button button");
 const compilerOutputButton: HTMLButtonElement = document.querySelector("#compiler-output-button button");
 
-let ScriptRunning: boolean;
+// Global web worker that will run the code
+let worker: Worker | undefined = undefined;
 
-const localStorageCodeInput = localStorage.getItem("kipper-code-editor-content");
-if (localStorageCodeInput != undefined) {
-  codeInput.value = localStorageCodeInput;
-  editorUpdate(localStorageCodeInput);
-} else {
-  codeInput.value = "";
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function runCode() {
-  console.log("Running Code!");
-  ScriptRunning = true;
+/**
+ * @summary Runs the code using a web worker and writes the console output to the "virtual" terminal.
+ */
+function runCode(): void {
   runCodeListItem.innerHTML = `<button>Stop</button>`;
   runCodeButton = document.querySelector("#run-code-list-item button");
   runCodeButton.addEventListener("click", stopCode);
 
-  // TODO: Actually run the Code
+  if (window.Worker) {
+    // If there is a program running at the moment, stop execution and run the most recent code.
+    if (worker !== undefined) {
+      stopCode();
+    }
+
+    // Clear the console
+    clearConsoleOutput();
+    clearCompilerOutput();
+
+    // Enable compiler logs
+    SwitchToCompilerOutput();
+
+    // Prepare the worker to run the code
+    // @ts-ignore
+    worker = new Worker(new URL('./compile/compile-worker.ts', import.meta.url));
+
+    // Event handler for the return. This imitates stdout.
+    let writeOntoConsoleOutput = false;
+    worker.onmessage = function(event) {
+      if (typeof event.data === 'string' || event.data instanceof String) {
+        if (writeOntoConsoleOutput) {
+          // Write stdout output
+          writeLineToConsoleOutput(event.data as string);
+        } else {
+          // Write compiler logs
+          writeLineToCompilerOutput(event.data as string);
+        }
+      } else if (typeof event.data === 'number' || event.data instanceof Number) {
+        if (writeOntoConsoleOutput) {
+          // Stop as this is the console exit status
+          writeLineToConsoleOutput(`\nFinished execution with exit code ${event.data}.`);
+          stopCode();
+        } else {
+          // Enable output to 'stdout'
+          SwitchToConsoleOutput();
+          writeOntoConsoleOutput = true;
+        }
+      } else {
+        console.error(`Invalid message from WebWorker: ${event.data}`);
+      }
+    }
+
+    // Post the message to tell the worker to process the code
+    const currentCode = codeTextArea.value;
+    worker.postMessage(currentCode);
+  } else {
+    alert("Your browser does not support web-workers! Aborting operation.");
+  }
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function stopCode() {
-  console.log("Stopped!");
-  ScriptRunning = false;
+function stopCode(): void {
   runCodeListItem.innerHTML = `<button>Run</button>`;
   runCodeButton = document.querySelector("#run-code-list-item button");
   runCodeButton.addEventListener("click", runCode);
 
-  // TODO: Actually stop the Code
+  if (window.Worker) {
+    // If there is no current execution, return.
+    if (worker === undefined) {
+      return;
+    }
+
+    // Terminate the worker
+    worker.terminate();
+    worker = undefined;
+  } else {
+    alert("Your browser does not support web-workers! Aborting operation.");
+  }
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function clearContent() {
+function clearContent(): void {
   console.log("Code Cleared!");
-  codeInput.value = "";
-  codeInputResult.innerHTML = "";
+  codeTextArea.value = "";
+  codeTextAreaResult.innerHTML = "";
+  localStorage.setItem(localStorageIdentifier, "");
   textSavingState.innerHTML = `<p class="gray-text">Code cleared!</p>`;
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function copyContent() {
+function copyContent(): void {
   console.log("Code Copied!");
-  navigator.clipboard.writeText(codeInput.value).then(() => {
+  navigator.clipboard.writeText(codeTextArea.value).then(() => {
     textSavingState.innerHTML = `<p class="gray-text">Code copied!</p>`;
   });
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
+let consoleOutput = "";
+let compilerOutput = "";
+
 function SwitchToConsoleOutput() {
-  // this will depend on the future implementation. Stop Development on this
+  // Change styling
+  compilerOutputButton.style.borderBottom = "2px solid var(--scheme-gray)";
+  consoleOutputButton.style.borderBottom = "3px solid var(--scheme-primary)";
+
+  writeConsoleResultAndHighlight(consoleOutput);
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
 function SwitchToCompilerOutput() {
-  // this will depend on the future implementation. Stop Development on this
+  // Change styling
+  consoleOutputButton.style.borderBottom = "2px solid var(--scheme-gray)";
+  compilerOutputButton.style.borderBottom = "3px solid var(--scheme-primary)";
+
+  writeConsoleResultAndHighlight(compilerOutput);
 }
 
 // if the input is not empty, signalise that code was restored
-if (codeInput.value != "")
+if (codeTextArea.value.trim() !== "")
   textSavingState.innerHTML = `<p class="gray-text">Welcome back! We restored the code of your last session for you :)</p>`;
 else
   textSavingState.innerHTML = `<p class="gray-text">Start typing! We will save your changes while you are typing!</p>`;
@@ -94,12 +155,30 @@ clearContentButton.addEventListener("click", clearContent);
 consoleOutputButton.addEventListener("click", SwitchToConsoleOutput);
 compilerOutputButton.addEventListener("click", SwitchToCompilerOutput);
 
+codeTextArea.addEventListener("input", event => {
+  const givenTextArea: HTMLTextAreaElement = event.target as HTMLTextAreaElement;
+  writeEditorResultAndHighlight(givenTextArea.value);
+})
+
+codeTextArea.addEventListener("scroll", () => {
+  syncTextAreaSizeAndScroll();
+})
+
+codeTextArea.addEventListener('keydown', (event) => {
+  checkForTab(event);
+});
+
+// Properly configure the sizes of the items in the browser window. This should set every item relative to the maximum
+// possible space available.
+window.addEventListener('DOMContentLoaded', setEditorAndConsoleSizes);
+window.addEventListener('resize', setEditorAndConsoleSizes);
+
 // runtime variable for the writing event listener
 let cancel;
 let spinning: boolean;
 
 // adding keyup listener
-codeInput.addEventListener("keyup", event => {
+codeTextArea.addEventListener("keyup", event => {
   // if cancel exists / is active -> clear timeout
   if (cancel)
     clearTimeout(cancel);
@@ -109,7 +188,7 @@ codeInput.addEventListener("keyup", event => {
   // only done when the user finished typing!
   cancel = setTimeout(() => {
     const givenTextArea: HTMLTextAreaElement = event.target as HTMLTextAreaElement;
-    localStorage.setItem("kipper-code-editor-content", givenTextArea.value);
+    localStorage.setItem(localStorageIdentifier, givenTextArea.value);
 
     spinning = false;
     textSavingState.innerHTML = `<p class="gray-text">Code Saved!</p>`;
@@ -137,11 +216,37 @@ codeInput.addEventListener("keyup", event => {
   }
 });
 
+// Initialise the codeInput
+(() => {
+  const localStorageCodeInput = localStorage.getItem(localStorageIdentifier);
+  if (localStorageCodeInput != undefined) {
+    codeTextArea.value = localStorageCodeInput;
+    writeEditorResultAndHighlight(localStorageCodeInput);
+  } else {
+    codeTextArea.value = "";
+  }
+})();
+
+// Initialise the default console output
+(() => {
+  const welcomeMessage: Array<string> = [
+    "Welcome to the Kipper Playground!\n",
+    "Try out your first program by writing:\n",
+    "  call print(\"Hello world\");"
+  ];
+
+  // Write to the console
+  for (const msg of welcomeMessage) {
+    writeLineToConsoleOutput(msg);
+  }
+  SwitchToConsoleOutput();
+})();
+
 /**
  * Editor-Update, which allows for syntax highlighting
  * @param value The value the element was updated to
  */
-function editorUpdate(value: string) {
+function writeEditorResultAndHighlight(value: string): void {
   // If the last character is a newline character
   // Add a placeholder space character to the final line
   if (value[value.length - 1] == "\n") {
@@ -149,24 +254,111 @@ function editorUpdate(value: string) {
   }
 
   // Write results to the original 'codeInput' <textarea> and syntax-highlighted result
-  codeInput.innerText = value;
-  codeInputResult.innerHTML = value
+  codeTextAreaResult.innerHTML = value
     .replace(new RegExp("&", "g"), "&")
     .replace(new RegExp("<", "g"), "<"); // Allow newlines
 
   // @ts-ignore
   // Prism should be imported
-  Prism.highlightElement(codeInputResult);
+  Prism.highlightElement(codeTextAreaResult);
+
+  // Sync formatting
+  syncTextAreaSizeAndScroll();
+}
+
+function checkForTab(event) {
+  const element = codeTextArea;
+  const code = element.value;
+  if(event.key == "Tab") {
+    event.preventDefault();
+
+    const beforeTab = code.slice(0, element.selectionStart);
+    const afterTab = code.slice(element.selectionEnd, element.value.length);
+
+    // where cursor moves after tab - moving forward by 1 char to after tab
+    const cursorPos = element.selectionEnd + 1;
+
+    // Add tab char
+    element.value = beforeTab + "\t" + afterTab;
+
+    // Move cursor
+    element.selectionStart = cursorPos;
+    element.selectionEnd = cursorPos;
+  }
 }
 
 /**
- * Syncs the scrolling for both <textarea> and codeInputResult
+ * Syncs the scrolling for both <textarea> and codeInputResult.
  */
-function syncScroll() {
+function syncTextAreaSizeAndScroll(): void {
   /* Scroll result to scroll coords of event - sync with textarea */
 
   // Get and set x and y
-  codeInputResult.scrollTop = codeInput.scrollTop;
-  codeInputResult.scrollLeft = codeInput.scrollLeft;
+  codeTextAreaResultWrapper.scrollTop = codeTextArea.scrollTop;
+  codeTextAreaResultWrapper.scrollLeft = codeTextArea.scrollLeft;
 }
 
+/**
+ * Write the passed text onto the console and applies syntax highlighting.
+ * @param value The text to write.
+ */
+function writeConsoleResultAndHighlight(value: string): void {
+  // If the last character is a newline character
+  // Add a placeholder space character to the final line
+  if (value[value.length - 1] == "\n") {
+    value += " ";
+  }
+
+  // Write content to the console
+  shellOutputResult.innerHTML = value
+    .replace(new RegExp("&", "g"), "&")
+    .replace(new RegExp("<", "g"), "<"); // Allow newlines
+
+  // @ts-ignore
+  // Prism should be imported
+  Prism.highlightElement(shellOutputResult);
+}
+
+/**
+ * Appends a new line to the console output and applies syntax highlighting.
+ * @param value The line to add.
+ */
+function writeLineToConsoleOutput(value: string): void {
+  consoleOutput += value + '\n';
+  writeConsoleResultAndHighlight(consoleOutput);
+}
+
+function writeLineToCompilerOutput(value: string): void {
+  compilerOutput += value + '\n';
+  writeConsoleResultAndHighlight(compilerOutput);
+}
+
+/**
+ * Clears the content of the console.
+ */
+function clearConsoleOutput(): void {
+  consoleOutput = "";
+  writeConsoleResultAndHighlight(compilerOutput);
+}
+
+function clearCompilerOutput(): void {
+  compilerOutput = "";
+  writeConsoleResultAndHighlight("");
+}
+
+/**
+ * Fixes the sizes of the code editor
+ */
+function setEditorAndConsoleSizes(): void {
+  const rem = parseFloat(getComputedStyle(document.documentElement).fontSize);
+
+  // Set editor size. Subtracts -2rem due to an inner padding of 1rem
+  codeTextAreaResultWrapper.style.height = `${codeEditor.clientHeight - 2 * rem}px`;
+  codeTextAreaResultWrapper.style.width = `${codeEditor.clientWidth - 2 * rem}px`;
+  codeTextArea.style.height = `${codeEditor.clientHeight - 2 * rem}px`;
+  codeTextArea.style.width = `${codeEditor.clientWidth - 2 * rem}px`;
+
+  // Set console size. Subtracts -2rem due to an inner padding of 1rem
+  shellOutputResult.style.height = `${shellOutput.clientHeight - 2 * rem}px`;
+  shellOutputResult.style.width = `${shellOutput.clientWidth - 2 * rem}px`;
+}